Workaround strict version checking in pcscd
Accepts 4.3 and 4.4

+++ doc.go
@@ -31,7 +31,7 @@
 package pcsc
 
 const (
-	SCardSuccess                   = 0x00000000 /* No error was encountered. */
+	SCardSuccess = 0x00000000 /* No error was encountered. */
 
 	AutoAllocate  = -1     /* see SCardFreeMemory() */
 	ScopeUser     = 0x0000 /* Scope in user space */
@@ -94,6 +94,6 @@ const (
 
 // Protocol information
 const (
-	ProtocolVersionMajor = 4 /* IPC major */
-	ProtocolVersionMinor = 3 /* IPC minor */
+	ProtocolVersionMajor = uint32(4) /* IPC major */
+	ProtocolVersionMinor = uint32(3) /* IPC minor */
 )
